requirements.txt
dash>=2.17.1
plotly>=5.23.0
psycopg2>=2.9.9
matplotlib>=3.9.1
reportlab>=4.2.2
kaleido>=0.2.1
python-dotenv